pyproject.toml
[tool.poetry]
name = "langchain_snl"
version = "0.1.0"
description = ""
authors = ["Danuk <danuk77@live.co.uk>"]
readme = "README.md"

[tool.poetry.scripts]
example = "langchain_snl:example"
chat = "langchain_snl:chat_model"
chat_template = "langchain_snl:chat_model_with_prompt_template"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.1"
langchain-anthropic = "^0.2.1"
langchain-core = "^0.3.6"
boto3 = "^1.35.29"
python-dotenv = "^1.0.1"
langchain-aws = "^0.2.1"
langgraph = "^0.2.28"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
